test: centralize the mock HTTP server in tests/http_mock (sonar duplication)
The fixed-body mock-server boilerplate was copied across test_wcs.py, test_wfs.py,
and test_oapif.py, tripping SonarCloud's new-code duplication gate. Extract one
make_fixed_body_server into tests/http_mock.py and reduce all three _make_server
copies to thin wrappers, dropping their now-unused http.server/socketserver/
threading/Counter imports.
